bugfix: Use JSON::PP which is thread-safe.
The following error was (sporadically) encountered on Raspbian GNU/Linux 9 (stretch):
"JSON text must be an object or array (but found number, string, true, false or null, use allow_nonref to allow this)"

The interpreter was "perl v5.24.1 built for arm-linux-gnueabihf-thread-multi-64int".

It turned out that the default module JSON::XS is not thread-safe.

Note: Another thread-safe altnernative is Cpanel::JSON::XS.